+++ tests/tuning_file.py
@@ -1,12 +1,12 @@
 #!/usr/bin/python3
 from picamera2 import CameraConfig, Picamera2
-
+from picamera2.tuning import load_tuning_file, find_tuning_algo
 # Here we load up the tuning for the HQ cam and alter the default exposure profile.
 # For more information on what can be changed, see chapter 5 in
 # https://datasheets.raspberrypi.com/camera/raspberry-pi-camera-guide.pdf
 
-tuning = Picamera2.load_tuning_file("imx477.json")
-algo = Picamera2.find_tuning_algo(tuning, "rpi.agc")
+tuning = load_tuning_file("imx477.json")
+algo = find_tuning_algo(tuning, "rpi.agc")
 algo["exposure_modes"]["normal"] = {"shutter": [100, 66666], "gain": [1.0, 8.0]}
 
 camera = Picamera2(tuning=tuning)
